# server/api/openapi.yaml
openapi: 3.0.1
info:
  title: API set for the To Do Manager web app [DSP version] 
  version: "1.0"
servers:
- url: /
...
paths:
...
  /tasks/{TaskId}/assignedTo:
    get:
      tags:
      - assigned tasks
      security:
        - cookieAuth: []
      summary: Retrieve the list of all the assignes of that task
      operationId: tasksTaskIdAssignedToGET
      parameters:
      - name: TaskId
        in: path
        description: Task id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  _links:
                    type: object 
                example:
                  users: [
                    {
                      "name": "name",
                      "id": "id",
                      "email": "example@exmple.com"
                    }
                  ]
                  _links: 
                    operation1: "url"
                    operation2: "url"
        "401":
          description: Unauthorized
        "404":
          description: Task Not Found
        "503":
          description: Internal Server Error
      x-swagger-router-controller: AssignedTasks
  /tasks/{TaskId}/assignedTo/{UserId}:
    put:
      tags:
      - assigned tasks
      security:
        - cookieAuth: []
      summary: "Assign an [already existing] user to the assignees user list of that\
        \ task"
      operationId: tasksTaskIdAssignedToUserIdPUT
      parameters:
      - name: TaskId
        in: path
        description: Task id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: UserId
        in: path
        description: User id to update inside the assigned user list
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content (successfully created)
        "400":
          description: Bad Request
        "401":
          description: Unauthorized  
        "403":
          description: Forbbiden  
        "404":
          description: Task/User not found
        "503":
          description: Internal Server Error
      x-swagger-router-controller: AssignedTasks
    delete:
      tags:
      - assigned tasks
      security:
        - cookieAuth: []
      summary: Remove an user from the assignees user list of that task
      operationId: tasksTaskIdAssignedToUserIdDELETE
      parameters:
      - name: TaskId
        in: path
        description: Task id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: UserId
        in: path
        description: User id to remove inside the assignees user list
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized  
        "403":
          description: Forbbiden  
        "404":
          description: Task not found
        "503":
          description: Internal Server Error
      x-swagger-router-controller: AssignedTasks
  /tasks/{TaskId}/images:
    post:
      tags:
      - task images
      security:
        - cookieAuth: []
      summary: "Add a new image to a speficic task (Image ID is automatically generated by the server)"
      operationId: tasksTaskIdImagesPOST
      requestBody:
        description: image to assign to the task
        content:
          multipart/form-data:
            schema:
              type: object
              properties: # Request parts
                 image:
                  type: string
                  format: binary
            encoding: 
                image: 
                  contentType: image/png, image/jpg, image/gif 
      responses:
        "201":
          description: Created
          content:
            multipart/form-data:
              schema:
                type: object
                properties: # Request parts
                  image:
                      type: string
                      format: binary
              encoding: 
                  image: 
                    contentType: image/png, image/jpg, image/gif 
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '415':
          description: Unsupported Media Type
        "503":
          description: Internal Server Error
      x-swagger-router-controller: Tasks    
  /tasks/{TaskId}/images/{ImageId}:
    get:
      tags:
      - task images
      security:
        - cookieAuth: []
      summary: Retrieve the requested image associated to the speficic task
      operationId: tasksTaskIdImagesImageIdGET
      parameters:
      - name: TaskId
        in: path
        description: Task id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: ImageId
        in: path
        description: Image id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Ok
          content:
            multipart/form-data:
              schema:
                type: object
                properties: # Request parts
                  image:
                      type: string
                      format: binary
              encoding: 
                  image: 
                    contentType: image/png, image/jpg, image/gif 
        "401":
          description: Unauthorized       
        "404":
          description: Not Found
        "503":
          description: Internal Server Error
      x-swagger-router-controller: Tasks 
    delete:
      tags:
      - task images
      security:
        - cookieAuth: []
      summary: Remove an user from the assignees user list of that task
      operationId: tasksTaskIdImagesImageIdDELETE
      parameters:
      - name: TaskId
        in: path
        description: Task id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: ImageId
        in: path
        description: Image id of the image that needs to be removed from the task speficic image list
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized  
        "403":
          description: Forbbiden  
        "404":
          description: Not found
        "503":
          description: Internal Server Error
      x-swagger-router-controller: AssignedTasks
...
  /login:
    post:
      tags:
      - login 
      summary: Perform the login process
      operationId: loginPOST
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "503":
          description: Internal Server Error
      x-swagger-router-controller: Login
  /logout:
    post:
      tags:
      - login 
      summary: Perform the login process
      operationId: logoutPOST
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "503":
          description: Internal Server Error
      x-swagger-router-controller: Login
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          maximum: 20
          minimum: 6
          type: string
      example:
        password: password
        name: name
        id: id
        email: ""
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid